+++ docs/pages/methods.js
@@ -8,7 +8,7 @@ import Param from 'components/param';
 import Example from 'components/example';
 import Code from 'components/code';
 import Paragraph from 'components/common/paragraph';
-import * as examples from 'examples/commonExamples';
+import * as examples from 'examples/code/commonExamples';
 import Link from 'components/common/link';
 import dataTypes from 'data/dataTypes';
 import {useIntl} from 'react-intl';
@@ -134,16 +134,22 @@ function Methods({} = {}) {
                                 <Param name={'destroy()'} id={'destroy'}>
                                     <Paragraph id={'apiDestroy'}/>
                                 </Param>
-                                <Param name={'update(newOpts)'} id={'update'}>
+                                <Param name={'update(newOpts, params?)'} id={'update'}>
                                     <Paragraph id={'apiUpdate'} />
                                     <Param.List nested>
                                         <Param name={'newOpts'} addId={false} type={'object'} definition={'apiUpdateNewOpts'} />
+                                        <Param name={'params.silent'} addId={false} type={'boolean'} definition={'apiUpdateSilent'} definitionValues={{
+                                            trueField
+                                        }} />
                                     </Param.List>
                                 </Param>
-                                <Param name={'setCurrentView(view)'} id={'setCurrentView'}>
+                                <Param name={'setCurrentView(view, params?)'} id={'setCurrentView'}>
                                     <Paragraph id={'apiSetCurrentView'} />
                                     <Param.List nested>
                                         <Param name={'view'} addId={false} type={'"days" | "months" | "years"'} definition={'apiSetCurrentViewView'} />
+                                        <Param name={'params.silent'} addId={false} type={'boolean'} definition={'apiSetCurrentViewSilent'} definitionValues={{
+                                            trueField
+                                        }} />
                                     </Param.List>
                                 </Param>
                                 <Param name={'setViewDate(date)'} id={'setViewDate'}>
@@ -179,6 +185,15 @@ function Methods({} = {}) {
                                         <Param addId={false} name={'date'} type={dataTypes.date} definition={'apiUpDate'} />
                                     </Param.List>
                                 </Param>
+                                <Param name={'getViewDates(view?)'} type={'(days|months|years) => Date[]'}>
+                                    <Paragraph id={'apiGetViewDates'} />
+                                </Param>
+                                <Param name={'disableDate(date)'} type={'((date: Date|string|number|Array<Date|string|number>)) => void'}>
+                                    <Paragraph id={'apiDisableDate'} />
+                                </Param>
+                                <Param name={'enableDate(date)'} type={'((date: Date|string|number|Array<Date|string|number>)) => void'}>
+                                    <Paragraph id={'apiEnabledDate'} />
+                                </Param>
                             </Param.List>
                         </Section>
 
@@ -204,6 +219,12 @@ function Methods({} = {}) {
                                 </Param>
                                 <Param name={'visible'} type={'boolean'}>
                                     <Paragraph id={'apiVisible'} />
+                                </Param>
+                                <Param name={'disabledDates'} type={'Set<string>'}>
+                                    <Paragraph id={'apiDisabledDates'} />
+                                </Param>
+                                <Param name={'isDestroyed'} type={'boolean'}>
+                                    <Paragraph id={'apiIsDestroyed'} values={{trueField}}  />
                                 </Param>
                             </Param.List>
                         </Section>